Consolidate badge logic and restore provider-state fallback for unknowns
Two related fixes:

1. Drift between sidebar and editor pills: the two views had separate
   buildBadges/buildStateBadge implementations that had diverged
   (sidebar 'Issue · open' vs editor 'Issue'). Move both into a new
   src/views/badges.ts shared module so they can never silently drift
   again. Both views now produce 'Issue' for the open/active state;
   the variant-driven color carries the open hint visually.

2. Unrecognized provider states were dropped entirely after the
   earlier 'two Review received pills' fix. Restore the raw provider
   state as a fallback meta-pill — but only when buildStateBadge
   didn't already render a semantic badge for it. This brings back
   visibility for custom states (e.g. ADO 'In Code Review') without
   reintroducing the duplication.

Both the editor header and the sidebar item card now render a
fallback pill from item.providerState (editor) /
item.providerStateFallback (sidebar) when present.
